remove backRelations

+++ src/mutations/delete.js
@@ -5,6 +5,10 @@ import type {
 } from '../utils/definitions.js'
 
 import {
+  isScalar
+} from '../utils/graphql.js'
+
+import {
   GraphQLNonNull,
   GraphQLID,
   GraphQLObjectType
@@ -40,12 +44,78 @@ export default function (
     mutateAndGetPayload: (node, { rootValue: { currentUser, backend, webhooksProcessor } }) => {
       node = convertInputFieldsToInternalIds(node, clientTypes[modelName].clientSchema)
 
-      return backend.deleteNode(modelName, node.id, clientTypes[modelName].clientSchema, currentUser)
-      .then((node) => {
-        webhooksProcessor.nodeDeleted(node, modelName)
-        return node
-      })
-      .then((node) => ({[getFieldNameFromModelName(modelName)]: node}))
+      function getBackRelationNodes (relationField, nodeToDelete) {
+        if (relationField.isList) {
+          return backend.allNodesByRelation(
+            modelName,
+            nodeToDelete.id,
+            relationField.fieldName,
+            null,
+            clientTypes[modelName].clientSchema,
+            currentUser)
+          .then((nodes) => nodes.filter((node) => node !== null))
+        } else {
+          if (!nodeToDelete[`${relationField.fieldName}Id`]) {
+            return Promise.resolve([])
+          }
+          return backend.node(
+            relationField.typeIdentifier,
+            nodeToDelete[`${relationField.fieldName}Id`],
+            clientTypes[relationField.typeIdentifier].clientSchema,
+            currentUser).then((node) => node ? [node] : [])
+        }
+      }
+
+      // todo: this disregards isRequired=true on related node
+      function setInlinedBackRelationsToNull (nodeToDelete) {
+        const relationFields = clientTypes[modelName].clientSchema.fields
+          .filter((field) => field.backRelationName)
+
+        if (relationFields.length === 0) {
+          return Promise.resolve()
+        }
+
+        return Promise.all(relationFields.map((field) =>
+          getBackRelationNodes(field, nodeToDelete)
+            .then((relationNodes) => {
+              return Promise.all(relationNodes.map((relationNode) => {
+                relationNode[`${field.backRelationName}Id`] = null
+                return backend.updateNode(
+                  field.typeIdentifier,
+                  relationNode.id,
+                  relationNode,
+                  clientTypes[field.typeIdentifier].clientSchema,
+                  currentUser)
+              }))
+            })))
+      }
+
+      return backend.node(
+          modelName,
+          node.id,
+          clientTypes[modelName].clientSchema,
+          currentUser)
+        .then((nodeToDelete) => {
+          if (nodeToDelete === null) {
+            return Promise.reject(`'${modelName}' with id '${node.id}' does not exist`)
+          }
+
+          // remove indexed and inlined relations to and from this node
+          return setInlinedBackRelationsToNull(nodeToDelete).then(() =>
+            backend.removeAllRelationsTo(modelName, node.id)
+            .then(() =>
+              Promise.all(clientTypes[modelName].clientSchema.fields
+                .filter((field) => field.isList && !isScalar(field.typeIdentifier))
+                .map((field) => backend.removeAllRelationsFrom(modelName, node.id, field.fieldName))
+                ).then(() =>
+                  backend.deleteNode(modelName, node.id, clientTypes[modelName].clientSchema, currentUser)
+                  .then((node) => {
+                    webhooksProcessor.nodeDeleted(node, modelName)
+                    return node
+                  })
+                  .then((node) => ({[getFieldNameFromModelName(modelName)]: node}))
+            )))
+        })
     }
   })
 }
